feat: fixed the code to download the full course content available
The downloading of the course is now working and logs are also enabled for review. In an initial
phase of logging, might change the logging behaviour as I think its too much bloated right now, as
the files increases the logs tends to be more. I will try to minimize the logging for its proper
usage. Right now just bear with me!

BREAKING CHANGE: The previus code had some incorrect import paths and now, its fixed. The
downloading is working as expected. The program now takes the course name as an argument via -course
parameter.